converted target to to_global in pores and throats
